Add KeysController
