+++ src/studio/src/designer/backend/Controllers/TextsController.cs
@@ -75,38 +75,6 @@ namespace Altinn.Studio.Designer.Controllers
         }
 
         /// <summary>
-        /// Endpoint for getting a list of all keys.
-        /// </summary>
-        /// <param name="org">Unique identifier of the organisation responsible for the app.</param>
-        /// <param name="repo">Application identifier which is unique within an organisation.</param>
-        /// <returns>List of keys</returns>
-        [HttpGet]
-        [Produces("application/json")]
-        [ProducesResponseType(StatusCodes.Status200OK)]
-        [ProducesResponseType(StatusCodes.Status404NotFound)]
-        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-        [Route("keys")]
-        public async Task<ActionResult<List<string>>> Get(string org, string repo)
-        {
-            string developer = AuthenticationHelper.GetDeveloperUserName(HttpContext);
-            IList<string> languages = _languagesService.GetLanguages(org, repo, developer);
-
-            try
-            {
-                List<string> keys = await _textsService.GetKeys(org, repo, developer, languages);
-                return Ok(keys);
-            }
-            catch (IOException)
-            {
-                return NotFound($"The texts files needed to get keys does not exist.");
-            }
-            catch (JsonException)
-            {
-                return new ObjectResult(new { errorMessage = $"The format of the file, .texts.json, that you tried to access might be invalid." }) { StatusCode = 500 };
-            }
-        }
-
-        /// <summary>
         /// Endpoint for updating a text file for a specific language.
         /// </summary>
         /// <param name="org">Unique identifier of the organisation responsible for the app.</param>
